restructure, get rid of diagram, fix .md file
