+++ README.md
@@ -2,23 +2,9 @@
 
 ## About
 
-Liam's Distributable Chat Server Project contains at least 2 pieces. The server and at least 1 receiver.  You start the server and any number of recivers, then you whatever you enter on one screen will appear on the other. 
-## Nhelo
+Liam's Distributable Chat Server Project is a simple CLI chat server that is cross platform.
+It allows 2 or more people to communicate from seperate computers 
 
-Nhelo in this rendition acts as a host, as well as another lnet, it can send data to lnet(s) and read it. It will come in a package with a server config. 
-## Lnet
+## Usage
 
-Liam net or Lnet is a receiver. It connects to the server (Nhelo) and sends data to nhelo it also recives messages from nhelo. 
-
-# More
-This is the start of a chat server project involving a fully fledged distributable program which allows you to create chat rooms with another person, or maybe more. This project will have 3 different versions:
-
-<ol>
-<li>LMCE-Liam's multi computer echo server</li>
-<li>LMCCS-Liam's multi computer chat server</li>
-<li>LDCSP-Liam's distributable chat server project</li>
-</ol>
-
-# images
-
-![graph](dia.png)
+You can get full usage by typing `ldcsp` or `ldcsp help` into the command prompt if correctly installed.
